[rename] coupons : handler 이동, exception 필드 public으로 변경 #15
