+++ coupon/src/main/java/com/haot/coupon/common/exceptions/CustomCouponException.java
@@ -7,5 +7,5 @@ import lombok.Getter;
 @Getter
 @AllArgsConstructor
 public class CustomCouponException extends RuntimeException{
-    ResCodeIfs resCode;
+    public ResCodeIfs resCode;
 }
