+++ coupon/src/main/java/com/haot/coupon/common/exceptions/handler/CustomCouponExceptionHandler.java
@@ -1,5 +1,6 @@
-package com.haot.coupon.common.exceptions;
+package com.haot.coupon.common.exceptions.handler;
 
+import com.haot.coupon.common.exceptions.CustomCouponException;
 import com.haot.coupon.common.response.ApiResponse;
 import com.haot.coupon.common.response.ResCodeIfs;
 import lombok.extern.slf4j.Slf4j;
